refactor(tauri): lift the card art cache out of the desktop shell
The cache, the `/scryfall-img/` route and the LAN listener lived in
`src-tauri`, so nothing in the workspace could reach them. A headless host on
somebody's Ubuntu box cannot serve card art from a Tauri crate, and holding
the art is the whole reason to have one: it is the thing every client
otherwise downloads separately.

`manabrew-art-cache` is that code with the Tauri removed. `download_all` takes
a progress callback where the command used `app.emit`, and `ArtServer::spawn`
already took its cache as an argument. The shell keeps the process-wide
instance, `init` from the app data dir, and the six commands, which is all it
ever had that was actually about Tauri.

No behaviour change: the same 16 tests moved with the code and pass unchanged.
