+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "manabrew-rs/crates/manabrew-hub",
     "manabrew-rs/crates/manabrew-protocol",
     "manabrew-rs/crates/manabrew-relay-protocol",
+    "manabrew-rs/crates/manabrew-art-cache",
     "manabrew-rs/crates/manabrew-net",
     "manabrew-rs/crates/manabrew-net-wasm",
     "manabrew-rs/crates/manabrew-server",
