+++ manabrew-rs/AGENTS.md
@@ -28,6 +28,7 @@ All under `manabrew-rs/crates/`. The relay, hub, and self-hosted node each have 
 | `manabrew-protocol`        | Engine-free DTOs shared by the clients: decks, game-domain types (`ManaColor`, `GameFormat`, `EngineKind`), prompts, display events, agent transport envelopes. Published. Holds no relay message and never depends on the relay crate.                                                                                                                                                                                                                                                                                                                                                 | —                                            |
 | `manabrew-relay-protocol`  | The relay wire protocol: `ClientMessage`/`ServerMessage`, the session handshake (`Authenticate`, `AuthResult`, `SessionTakenOver`), lobby management (`RoomInfo`, `PlayerInfo`, `DraftConfig`, `SealedConfig`), and `PROTOCOL_VERSION`. Depends on `manabrew-protocol` for `Deck` and the domain enums; nothing there depends back. **`PROTOCOL_VERSION` is this crate's major**, so a breaking wire change ships as a major release here, never as a separate constant. Owns the `gen-protocol` bin, which sits at the top of the protocol graph and emits TypeScript for both crates. | —                                            |
 | `manabrew-game-runtime`    | Hosts an N-player game session over a caller-supplied transport. `host_runtime::run_hosted_multiplayer_game` is the single source of truth for the multiplayer setup-and-run dance (Tauri uses mpsc, WASM uses SharedArrayBuffer — both supply their transport via closures).                                                                                                                                                                                                                                                                                                           | —                                            |
+| `manabrew-art-cache`       | Card art on disk, the `/scryfall-img/` route it backs, and the listener that serves it to a network. No Tauri: the desktop shell owns one instance and wraps it in commands, a headless host owns another.                                                                                                                                                                                                                                                                                                                                                                              | —                                            |
 | `manabrew-net`             | The direct data plane: iroh endpoint lifecycle, session framing, the per-seat QUIC channel, and the room-scoped iroh-gossip topic. Depends only on `manabrew-relay-protocol`, so the wire protocol stays transport-agnostic. See `docs/TRANSPORT.md`.                                                                                                                                                                                                                                                                                                                                   | —                                            |
 | `networking-tests`         | System-wide networking regression suite: spawns the real relay + node binaries and drives real websocket clients (production `manabot` for bot seats). One Given/When/Then per test; `yarn test:networking` locally, build-checks CI serialized. Custom `libtest-mimic` harness (`harness = false`): each test is a plain `async fn` that must be registered in `main()` — a bare `#[tokio::test]`/`#[test]` will compile but never run.                                                                                                                                                | —                                            |
 | `loadgen`                  | Load generator: drives real games against a relay so whatever hosts them can be measured under load. Claims hosted rooms by default; with `ROOM` it joins a named room instead, which covers a browser hosting a table. Real protocol, `manabot` decisions, bot seats. Env-configured; see its README.                                                                                                                                                                                                                                                                                  | —                                            |
