+++ ops/manifest.json
@@ -15,6 +15,7 @@
     "manabrew-engine": "0.8.2",
     "manabrew-engine-macros": "0.2.0",
     "manabrew-hub": "0.15.1",
+    "manabrew-art-cache": "0.1.0",
     "manabrew-net": "0.1.0",
     "manabrew-net-wasm": "0.1.0",
     "manabrew-protocol": "5.6.1",
