+++ src-tauri/AGENTS.md
@@ -16,8 +16,7 @@ Read first: `/AGENTS.md`.
 | `src/lan_discovery.rs`      | mDNS advertise/browse so a LAN room is found without typing an address (feature `forge-room`).                                                                                                                                                                                     |
 | `src/direct_seat.rs`        | This machine's **native** iroh endpoint for its own seat (feature `direct-seat`, implied by `forge-room`). `forge-room` also needs `self-hosted-node/iroh`, or a hosted room announces no endpoint and every LAN seat quietly relays; `local_relay.rs` asserts it at compile time. |
 | `src/card_db.rs`            | Card database loader (rkyv archive) backing `is_card_supported` and the limited stack.                                                                                                                                                                                             |
-| `src/image_cache.rs`        | Card art on disk plus the `/scryfall-img/` route it backs, and the `preseed_card_art` download.                                                                                                                                                                                    |
-| `src/art_lan_server.rs`     | The opt-in listener that serves that cache to the local network while a room is shared.                                                                                                                                                                                            |
+| `src/image_cache.rs`        | This machine's instance of `manabrew-art-cache`, and the commands the webview drives it with. The cache and its LAN listener live in that crate so a headless host can use both.                                                                                                   |
 | `src/limited_*.rs`          | Sealed / draft / winston / gauntlet / cube — the limited-format command surface and its session manager.                                                                                                                                                                           |
 | `tauri.conf.json`           | Bundle config (forge runtime + cardset archive) and the COOP/COEP headers that make WASM/SharedArrayBuffer work in the webview.                                                                                                                                                    |
